% File src/library/base/man/demereldist.Rd
     \name{demereldist}
     \docType{data}
     \alias{demereldist}
     \title{Example dataset for package Demerelate to calculate inter individual pairwise genetic relatedness.}
     \description{The dataset gives randomized relative positions for each individual in the dataset.}

     \usage{
     data(demereldist)
     }
     

\author{Philipp Kraemer, <philipp.kraemer@uni-oldenburg.de>}

     \seealso{
       \code{\link{inputformat}
	     \link{Demerelate}
	      }
     }
     \examples{

## Consult examples from Demerelate
    
     
     }
