% File src/library/base/man/demerelref.Rd
     \name{demerelref}
     \docType{data}
     \alias{demerelref}
     \title{Example dataset for package Demerelate to calculate inter individual pairwise genetic relatedness.}
     \description{
       The dataset gives randomized loci information of a population with 1000 individuals. Dataset is based on information of 8 diploid loci with total number in alleles indicated by column names, i.e. number of alleles = 5, 10, 15, 20, 25, 30, 35, 40.
     }

     \usage{
     data(demerelref)
     }
     

\author{Philipp Kraemer, <philipp.kraemer@uni-oldenburg.de>}

     \seealso{
       \code{
	     \link{Demerelate}
	     \link{demereldist}
	     \link{demerelpop}
	      }
     }
     \examples{

## Consult examples from demerelpop
    
     
     }
